07-05 beta1

+++ trunk/TDCGExplorer/TDCGExplorer/TAHDump.cs
@@ -40,13 +40,16 @@ namespace TDCGExplorer
         public static void ArcsDumpDirEntriesMain(string dir, ArcsDatabase db)
         {
             arcspath = dir;
-            // 存在フラグを全て落とす.
-            TDCGExplorer.SetToolTips("Setup database");
-            db.UpdateTahExistDown();
-            ArcsDumpDirEntries(dir, db);
-            // 存在しないtahファイルは消去する.
-            TDCGExplorer.SetToolTips("Deleting orphan records");
-            db.DeleteNoExistentTah();
+            if (Directory.Exists(dir) == true)
+            {
+                // 存在フラグを全て落とす.
+                TDCGExplorer.SetToolTips("Setup database");
+                db.UpdateTahExistDown();
+                ArcsDumpDirEntries(dir, db);
+                // 存在しないtahファイルは消去する.
+                TDCGExplorer.SetToolTips("Deleting orphan records");
+                db.DeleteNoExistentTah();
+            }
         }
 
         public static void ArcsDumpDirEntries(string dir,ArcsDatabase db)
@@ -138,7 +141,7 @@ namespace TDCGExplorer
 
         public static void ArcsDumpTahFilesEntries(ArcsDatabase db,ArcsTahEntry entry)
         {
-            string source = TDCGExplorer.GetSystemDatabase().arcs_path + "\\" + entry.path;
+            string source = Path.Combine(TDCGExplorer.GetSystemDatabase().arcs_path, entry.path);
 
             TAHFile tah = new TAHFile(source);
             try
@@ -215,14 +218,17 @@ namespace TDCGExplorer
         public static void ZipsDumpDirEntriesMain(string dir, ArcsDatabase db)
         {
             zipspath = dir;
-            zipcoderegexp = TDCGExplorer.GetSystemDatabase().zip_regexp;
-            // 存在フラグを全て落とす.
-            TDCGExplorer.SetToolTips("Setup database");
-            db.UpdateZipExistDown();
-            ZipsDumpDirEntries(dir, db);
-            // 存在しないtahファイルは消去する.
-            TDCGExplorer.SetToolTips("Deleting orphan records");
-            db.DeleteNoExistentZip();
+            if (Directory.Exists(dir) == true)
+            {
+                zipcoderegexp = TDCGExplorer.GetSystemDatabase().zip_regexp;
+                // 存在フラグを全て落とす.
+                TDCGExplorer.SetToolTips("Setup database");
+                db.UpdateZipExistDown();
+                ZipsDumpDirEntries(dir, db);
+                // 存在しないtahファイルは消去する.
+                TDCGExplorer.SetToolTips("Deleting orphan records");
+                db.DeleteNoExistentZip();
+            }
         }
 
         public static void ZipsDumpDirEntries(string dir, ArcsDatabase db)
@@ -307,19 +313,19 @@ namespace TDCGExplorer
                 case ".zip":
                     using (IArchive arc = new ZipArchive())
                     {
-                        DumpArcEntries(db, TDCGExplorer.GetSystemDatabase().zips_path + "\\" + entry.path, arc, entry.id);
+                        DumpArcEntries(db, Path.Combine(TDCGExplorer.GetSystemDatabase().zips_path, entry.path), arc, entry.id);
                     }
                     break;
                 case ".rar":
                     using (IArchive arc = new RarArchive())
                     {
-                        DumpArcEntries(db, TDCGExplorer.GetSystemDatabase().zips_path + "\\" + entry.path, arc, entry.id);
+                        DumpArcEntries(db, Path.Combine(TDCGExplorer.GetSystemDatabase().zips_path, entry.path), arc, entry.id);
                     }
                     break;
                 case ".lzh":
                     using (IArchive arc = new LzhArchive())
                     {
-                        DumpArcEntries(db, TDCGExplorer.GetSystemDatabase().zips_path + "\\" + entry.path, arc, entry.id);
+                        DumpArcEntries(db, Path.Combine(TDCGExplorer.GetSystemDatabase().zips_path, entry.path), arc, entry.id);
                     }
                     break;
                 default:
